+++ decider-bcl2fastq/src/main/java/ca/on/oicr/pde/deciders/Bcl2fastqDeciderCli.java
@@ -85,7 +85,7 @@ public class Bcl2fastqDeciderCli extends Plugin implements DeciderInterface {
     private final Bcl2fastqDecider decider;
     private final OptionSpec<Boolean> noLaneSplittingOpt;
     private final OptionSpec<Boolean> ignoreLaneSkipOpt;
-    private final OptionSpec<Boolean> doNotProvisionUndeterminedOpt;
+    private final OptionSpec<Boolean> provisionUndeterminedOpt;
 
     public Bcl2fastqDeciderCli() {
         super();
@@ -154,9 +154,9 @@ public class Bcl2fastqDeciderCli extends Plugin implements DeciderInterface {
         ignoreLaneSkipOpt = parser.accepts("ignore-lane-skip",
                 "Ignore lane skip field and process lanes that are marked as skipped.")
                 .withOptionalArg().ofType(Boolean.class).defaultsTo(false);
-        doNotProvisionUndeterminedOpt = parser.accepts("do-not-provision-out-undetermined",
-                "Do not provision out undetermined fastqs.")
-                .withOptionalArg().ofType(Boolean.class).defaultsTo(false);
+        provisionUndeterminedOpt = parser.accepts("provision-out-undetermined",
+                "Provision out undetermined fastqs.")
+                .withOptionalArg().ofType(Boolean.class).defaultsTo(true);
 
         //output options
         outputPathOpt = parser.accepts("output-path",
@@ -303,7 +303,7 @@ public class Bcl2fastqDeciderCli extends Plugin implements DeciderInterface {
         decider.setLaunchMax(options.valueOf(launchMaxOpt));
         decider.setNoLaneSplittingMode(getBooleanFlagOrArgValue(noLaneSplittingOpt));
         decider.setIgnoreLaneSkip(getBooleanFlagOrArgValue(ignoreLaneSkipOpt));
-        decider.setDoNotProvisionOutUndetermined(getBooleanFlagOrArgValue(doNotProvisionUndeterminedOpt));
+        decider.setProvisionOutUndetermined(getBooleanFlagOrArgValue(provisionUndeterminedOpt));
 
         decider.setOutputPath(options.valueOf(outputPathOpt).endsWith("/") ? options.valueOf(outputPathOpt) : options.valueOf(outputPathOpt) + "/");
         decider.setOutputFolder(options.valueOf(outputFolderOpt));