+++ decider-bcl2fastq/src/main/java/ca/on/oicr/pde/deciders/handlers/Bcl2Fastq_2_7_1_Handler.java
@@ -19,8 +19,8 @@ public class Bcl2Fastq_2_7_1_Handler extends Bcl2FastqHandler {
         if (data.getNoLaneSplitting()) {
             workflowRun.addError(Bcl2Fastq_2_7_1_Handler.class.getSimpleName() + " does not support no-lane-splitting");
         }
-        if (data.getDoNotProvisionOutUndetermined()) {
-            workflowRun.addError(Bcl2Fastq_2_7_1_Handler.class.getSimpleName() + " does not support do-not-provision-out-undetermined");
+        if (!data.getProvisionOutUndetermined()) {
+            workflowRun.addError(Bcl2Fastq_2_7_1_Handler.class.getSimpleName() + " does not support disabling provision-out-undetermined");
         }
     }