+++ decider-bcl2fastq/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.3 - 2019-01-??
 - [GP-1211](https://jira.oicr.on.ca/browse/GP-1121) - Support for single end sequencing
+- [GP-1900](https://jira.oicr.on.ca/browse/GP-1900) - bcl2fastq option to not provision out undetermined fastqs
 ## 1.2.4 - 2018-12-04
 - [GP-1883](https://jira.oicr.on.ca/browse/GP-1883) - Only retrieve analysis records from "check-wf-accessions"
 ## 1.2.3 - 2018-11-20